+++ libs/dot-components/src/lib/ctm/progression-board/ProgressionBoardInterfaces.tsx
@@ -78,6 +78,7 @@ export interface ProgressionBoardProps {
 }
 
 export interface SwimLaneProps {
+  className: string;
   package: SwimLanepkg;
   selectWorkitemProps: {
     selectWorkitem: (id) => void;